[Design] Redraw circle when the window size changes
