+++ assets/css/player.scss
@@ -61,7 +61,7 @@
 @include media-breakpoint-up(md) {
 
     #player {
-        width: 375px;
+        width: 330px;
     }
     
     #skip-back {
