+++ assets/js/player.js
@@ -1,77 +1,29 @@
 import Amplitude from 'amplitudejs/dist/amplitude';
+import ProgressCircle from './progress_circle';
 
 const player = document.querySelector("#player");
 
 if (player) {
 
     window.onload = function() {
-
+        ProgressCircle.init();
         const dataUrl = player.getAttribute("data-url");
-
-        // based on https://github.com/rendro/easy-pie-chart
-        const drawCircle = (ctx, color, radius, lineWidth, percent) => {
-            percent = Math.min(Math.max(0, percent || 1), 1);
-            ctx.beginPath();
-            ctx.arc(0, 0, radius, 0, Math.PI * 2 * percent, false);
-            ctx.strokeStyle = color;
-            ctx.lineCap = 'round';
-            ctx.lineWidth = lineWidth;
-            ctx.stroke();
-        };
+        
+        const updateProgressCircle = () => {
+            let percentage = Amplitude.getSongPlayedPercentage();
     
-        // settings for circle:
-        const progressbar = document.getElementById('progressbar');
-        const size = progressbar.offsetWidth;
-        const lineWidth = 30;
-        const canvas = document.createElement('canvas');
-        const ctx = canvas.getContext('2d');
-    
-        // increase the size of the canvas by the devicePixelRatio to make sure it's sharp:
-        canvas.width = canvas.height = size * window.devicePixelRatio;
-        // the canvas element must have the original size: (i.e. we'll put a larger canvas into a smaller element)
-        canvas.style.height = canvas.style.width = size + "px";
-    
-        const radius = (canvas.width - lineWidth) / 2;
-    
-        // append canvas:
-        progressbar.appendChild(canvas);
-    
-        // set translation to move circle to center:
-        ctx.translate(canvas.width / 2, canvas.width / 2);
-        ctx.rotate((-1 / 2) * Math.PI);
-        const translation = -1 * canvas.width / 2;
-    
-        const drawProgressCircle = (percentage) => {
-            if (percentage == 0) {
-                percentage = 0.1;
+            if( isNaN( percentage ) ){
+                percentage = 0;
             }
-    
-            // clear for redraw:
-            ctx.clearRect(translation, translation, canvas.width, canvas.height);
-    
-            // redraw:
-            drawCircle(ctx, '#efefef', radius, lineWidth, 100 / 100);
-            drawCircle(ctx, '#00456f', radius, lineWidth, percentage / 100);
-        }
-    
+            ProgressCircle.drawProgressCircle(percentage);
+        };
     
         // start amplitude player:
         Amplitude.init({
             "callbacks": {
-                timeupdate: function(){
-                    let percentage = Amplitude.getSongPlayedPercentage();
-    
-                    if( isNaN( percentage ) ){
-                        percentage = 0;
-                    }
-                    drawProgressCircle(percentage);
-                }
+                timeupdate: updateProgressCircle
             },
-            songs: [
-                {
-                    "url": dataUrl
-                }
-            ],
+            songs: [{"url": dataUrl}],
             playback_speed: 1.0
         });
     
@@ -82,7 +34,6 @@ if (player) {
             const skipBackSeconds = skipBackButton.getAttribute("data-skip-back-seconds");
             Amplitude.skipTo(Math.max(0, seconds - skipBackSeconds), 0);
         })
-    
     
         const playbackSpeedButton = document.getElementById("playback-speed");
     
@@ -106,6 +57,12 @@ if (player) {
         })
     
         // start with an empty circle:
-        drawProgressCircle(0);
+        ProgressCircle.drawProgressCircle(0);
+
+        // redraw on size change:
+        window.addEventListener("resize", (event) => {
+            ProgressCircle.init();
+            updateProgressCircle();
+        });
     }
 }
