+++ lib/qrstorage_web/templates/qr_code/show.html.heex
@@ -14,7 +14,6 @@
   <strong>Link</strong><br/>
   <p><%= @qr_code.text %></p>
 <% end %>
-<br />
 
 <%= if @qr_code.content_type == :audio do %>
   <div class="row">
